fix window

+++ Compiler/link/link.h
@@ -5,5 +5,5 @@ int initWindow ();
 int closeWindow ();
 
 // Interface for this task
-int putPixel (int x , int y , float heat);
+int putPixel (int x , int y , short heat);
 int flush ();
