+++ Compiler/math/math.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "../link/link.h"
 
-int initPixels (float pixels[SzW][SzH]);
-int updatePixels (float pixels[SzW][SzH]);
+int initPixels (short pixels[SzW][SzH]);
+int updatePixels (short pixels[SzW][SzH]);
 
 void check_impl (int err_code , const char* file , const unsigned line);
 #define check(statement) check_impl(statement, __FILE__, __LINE__);
